Space characters output in SWF is wrong unicode char 0xe000 instead of 0x0020
Using poppler, avoid checking if a character is an space relying on advance,
as we don't have such an information. Using that check, we were not detecting
the space as such, and were adding a new space character to the GfxFont. So
all existing spaces were wrongly converted to 0xe000.

https://forge.igalia.com/issues/894