pdf2swf does not add white background on bitmap and poly2bitmap output modes
bitmap and poly2bitmap where not adding a white background to the document
nor adding an option to allow this as a parameter, as vector output does. So
I added support for a new parameter flag "transparent", that fills the
output with a white background in case it is not provided.

https://forge.igalia.com/issues/905